Add a mini shell (#54)
* Set up build for minimal entry

* Adapt hello.c to launch cosh

* Add elementary repl

* st_ino should actually be unsigned short

* Add fstat operation for chardevs

* Add _execve to get it to link

* Actually build cosh

* (temporary) Changes for debugging (revert this later?)

* Refresh TLB on allocation and deallocation by cr3

How the f*** this was the bug that kept me up all night, I do not know. I'm even more distraught by the fact that adding two lines seems to have solved this god-damned issue.

* Revert "(temporary) Changes for debugging (revert this later?)"

This reverts commit 318f6bfd40dd00a3f283f6e82d0d0eff1f108f3d.

* Implement a base gen_argv

* Parse backslashed characters in doublequoted args

* Add *argc parameter

* Set up basic command dispatch with exit

* Add cd and chdir

* Add pwd

* Add ls

* Remove execve definition causing build issues

* Allow execing other binaries

* Add 'true' and 'false' keywords

* Add echo

* Add mkdir

* Add stat

* Add getpid

* Add touch

* Strip everything after the first valid #

* Parse semicolons and `$$`, `$?`

* Add eval

* Move strip_comment call to within read_line

* Add 'source' call

* Add support for elementary conditionals `&&` and `||`

* Add test

* Update kernel internal open flags and touch implementation

* Source argv[1] if it is passed

* Remove unused include

* Set PATH environment variable when launching cosh initially

* Add basic diagnosis script which can be run with cosh

* Update the readme

+++ kernel/include/kernel/fs/stat.h
@@ -9,7 +9,7 @@ typedef struct {
 
 typedef struct {
 	short		   st_dev;
-	uint64_t	   st_ino;
+	unsigned short st_ino;
 	uint32_t	   st_mode;
 	unsigned short st_nlink;
 	unsigned short st_uid;
